+++ projects/swagger-schematics/interfaces/version_3_1/response.interface.ts
@@ -2,6 +2,7 @@ import { THttpStatusCode } from '../http-status-code.enum';
 import { IContent } from './content.interface';
 import { TSchema } from './swagger.interface';
 import { IRef } from './ref.interface';
+import { IExample } from './params.interface';
 
 export interface IHeader {
     description?: string;
@@ -11,16 +12,16 @@ export interface IHeader {
     style?: 'simple';
     explode?: boolean;
     schema?: TSchema;
-    example?: any;
-    examples?: Record<string, any>;
+    example?: unknown;
+    examples?: Record<string, IExample | IRef>;
     content?: IContent;
 }
 
 export interface ILink {
     operationRef?: string;
     operationId?: string;
-    parameters?: Record<string, any>;
-    requestBody?: any;
+    parameters?: Record<string, unknown>;
+    requestBody?: unknown;
     description?: string;
     server?: { url: string; description?: string };
 }